Add files via upload
This project was undertaken with a dual motivation: to gain insights into historical events and to demonstrate data analysis and machine learning techniques. The dataset used in this analysis is derived from the Titanic passenger list, which includes information about passengers such as age, gender, ticket class, and whether they survived or not. This dataset, 'train.csv', is available in the 'data' directory and serves as the foundation for our analysis and predictions.